Convert service functions to TS
